transformer/requirements.txt
psycopg2-binary==2.9.9
flask==3.0.0
flask-cors==4.0.0



